SPDE CLV_SIR Simulation
This code contains the simulation for the spatio-temporal (SPDE) form of the CLV_SIRE model solution, considering both spatial movements and noise effects. This code considers both white noise and Levy jumps. This code also considers two population sets, one is non-disease-carrying initially, and the other is disease-carrying initially. The numerical method follows the discretization as a central finite difference method for the Laplacian associated with the diffusion, the upwind method to the advection, and the Euler-Maruyama method for the noise settings.